src/python/requirements.txt
PyGithub==1.47
jinja2==2.11.3